feat: section layout
