+++ vue/src/board/puzzle/ChartPuzzle.vue
@@ -1,7 +1,5 @@
 <template>
-  <div>
-    {{ chart.type }}
-  </div>
+  <div>{{ chart.type }}</div>
 </template>
 
 <script setup>
@@ -10,7 +8,7 @@
  * @file: ChartPuzzle.vue
  * @since: 2024-07-14 20:57:49
  **/
-// import charts from '@swanlab-vue/board/charts'
+import charts from '@swanlab-vue/board/charts'
 /**
  * @typedef {Object} ChartPuzzleProps
  * @property {Chart} chart - 图表配置
