+++ vue/src/board/puzzle/layout/MobileLayout.vue
@@ -1,4 +1,6 @@
-<template></template>
+<template>
+  <div>{{ charts }}</div>
+</template>
 
 <script setup>
 /**
@@ -6,6 +8,7 @@
  * @file: MobileLayout.vue
  * @since: 2024-07-15 17:09:47
  **/
+const props = defineProps(['charts'])
 </script>
 
 <style lang="scss" scoped></style>
