+++ vue/src/board/puzzle/layout/StandardLayout.vue
@@ -1,4 +1,16 @@
-<template></template>
+<template>
+  <Row :gutter="[24, 24]">
+    <Col
+      v-for="chart in charts"
+      :key="chart.index"
+      :md="24"
+      :lg="isMedia(chart.type) ? 24 : 12"
+      :xl="isMedia(chart.type) ? 24 : 8"
+    >
+      <ChartPuzzle :chart="chart" class="border" />
+    </Col>
+  </Row>
+</template>
 
 <script setup>
 /**
@@ -6,6 +18,13 @@
  * @file: StandardLayout.vue
  * @since: 2024-07-15 17:09:24
  **/
+import { Row, Col } from 'ant-design-vue'
+import ChartPuzzle from '../ChartPuzzle.vue'
+const props = defineProps(['charts'])
+
+const isMedia = (type) => {
+  return ['image', 'audio', 'text'].includes(type.toLowerCase())
+}
 </script>
 
 <style lang="scss" scoped></style>
